backend/pyproject.toml
[project]
name = "daidaex-backend"
version = "0.1.0"
description = "FastAPI backend for DaidaEx embeddings and search"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "sentence-transformers==3.0.1",
  "faiss-cpu==1.8.0",
  "numpy==1.26.4",
  "scikit-learn==1.5.1",
  "SQLAlchemy==2.0.35",
  "psycopg2-binary==2.9.9",
  "python-dotenv==1.0.1"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

